requirements.txt
amqp==2.1.1
anyjson==0.3.3
apiai==1.0.2
appnope==0.1.0
awscli==1.11.8
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.2
boto==2.43.0
botocore==1.4.65
celery==4.0.0
cffi==1.9.1
colorama==0.3.7
cryptography==1.6
cycler==0.10.0
dateutils==0.6.6
decorator==4.0.10
Django==1.10.2
django-extensions==1.7.4
djangorestframework==3.5.3
docopt==0.6.2
docutils==0.12
ecdsa==0.13
enum34==1.1.6
Fabric==1.12.0
fabutil==0.1.1
futures==3.0.5
gunicorn==19.6.0
idna==2.1
inflection==0.3.1
ipaddress==1.0.17
ipython==5.1.0
ipython-genutils==0.1.0
jmespath==0.9.0
kombu==4.0.0
more-itertools==2.3
ndg-httpsclient==0.4.2
numpy==1.11.2
pandas==0.19.1
paramiko==1.17.2
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
plotly==1.12.9
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.1.9
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
Pygments==2.1.3
pymongo==3.3.1
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-dateutil==2.5.3
python-http-client==2.2.1
pytz==2016.7
Quandl==3.0.1
records==0.4.3
requests==2.11.1
rsa==3.4.2
s3transfer==0.1.9
sendgrid==3.6.3
sendgrid-django==4.0.1
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
slackclient==1.0.2
SQLAlchemy==1.1.3
tablib==0.11.2
traitlets==4.3.1
vine==1.1.3
wcwidth==0.1.7
websocket-client==0.37.0
yahoo-finance==1.3.2
